requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3

# Database
sqlalchemy==2.0.36

# CrewAI + LLMs
crewai==0.86.0
crewai-tools==0.17.0
langchain-community==0.3.7
langchain-openai==0.2.8
ollama==0.4.0

# ML
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3

# Dashboard
streamlit==1.40.1
plotly==5.24.1

# Auth
PyJWT==2.10.0
bcrypt==4.2.1
python-multipart==0.0.17

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
